# PERF: remove categories