@phofl 2023-02-14
looks like we got it back to the original performance, thx